+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - JSON configuration support.
 - YAML configuration support.
 - HTTP handler for authentication.
+- Add token review Prometheus metrics.
 
 [unreleased]: https://github.com/slok/simple-ingress-external-auth/compare/v0.1.0...HEAD
 [v0.1.0]: https://github.com/slok/simple-ingress-external-auth/releases/tag/v0.1.0